+++ src/Firebase/Exception/QueryException.php
@@ -24,14 +24,14 @@ class QueryException extends \RuntimeException implements FirebaseException
     {
         $message = $e->getMessage();
 
-        if (stripos($message, 'index not defined') !== false) {
+        if (false !== stripos($message, 'index not defined')) {
             throw new IndexNotDefined($query, $message, $e->getCode(), $e);
         }
 
-        if (stripos($message, 'orderby must be defined') !== false) {
+        if (false !== stripos($message, 'orderby must be defined')) {
             $message = 'An order must be defined for a query when applying filters.'
                 .' You can use Query::orderByValue(), Query::orderByKey() or Query::orderByChild().';
-        } elseif (stripos($message, 'key index passed non') !== false) {
+        } elseif (false !== stripos($message, 'key index passed non')) {
             $message = 'The query is ordered by key, but you tried to filter it with a non-string value.';
         }