+++ src/Firebase/Factory.php
@@ -3,8 +3,13 @@
 namespace Kreait\Firebase;
 
 use Firebase\Auth\Token\Handler as TokenHandler;
+use GuzzleHttp\Client;
+use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7;
 use Kreait\Firebase;
+use Kreait\Firebase\Auth\CustomTokenGenerator;
+use Kreait\Firebase\Auth\IdTokenVerifier;
+use Kreait\Firebase\Http\Middleware;
 use Kreait\Firebase\ServiceAccount\Discoverer;
 use Psr\Http\Message\UriInterface;
 
@@ -30,11 +35,15 @@ final class Factory
      */
     private $serviceAccountDiscoverer;
 
+    /**
+     * @var string|null
+     */
+    private $apiKey;
+
     private static $databaseUriPattern = 'https://%s.firebaseio.com';
 
     /**
      * @deprecated 3.1 use {@see withServiceAccount()} instead
-     * @codeCoverageIgnore
      *
      * @param string $credentials Path to a credentials file
      *
@@ -52,10 +61,51 @@ final class Factory
         return $this->withServiceAccount(ServiceAccount::fromValue($credentials));
     }
 
+    /**
+     * @deprecated 3.2 use {@see withServiceAccountAndApiKey()} instead
+     *
+     * @param string $apiKey
+     *
+     * @return Factory
+     */
+    public function withApiKey(string $apiKey): self
+    {
+        trigger_error(
+            'This method is deprecated and will be removed in release 4.0 of this library.'
+            .' Use Firebase\Factory::withServiceAccountAndApiKey() instead.', E_USER_DEPRECATED
+        );
+
+        $factory = clone $this;
+        $factory->apiKey = $apiKey;
+
+        return $factory;
+    }
+
+    /**
+     * @deprecated 3.2 use {@see withServiceAccountAndApiKey()} instead
+     *
+     * @param ServiceAccount $serviceAccount
+     *
+     * @return self
+     */
     public function withServiceAccount(ServiceAccount $serviceAccount): self
+    {
+        trigger_error(
+            'This method is deprecated and will be removed in release 4.0 of this library.'
+            .' Use Firebase\Factory::withServiceAccountAndApiKey() instead.', E_USER_DEPRECATED
+        );
+
+        $factory = clone $this;
+        $factory->serviceAccount = $serviceAccount;
+
+        return $factory;
+    }
+
+    public function withServiceAccountAndApiKey(ServiceAccount $serviceAccount, string $apiKey): self
     {
         $factory = clone $this;
         $factory->serviceAccount = $serviceAccount;
+        $factory->apiKey = $apiKey;
 
         return $factory;
     }
@@ -76,8 +126,21 @@ final class Factory
         return $factory;
     }
 
+    /**
+     * @deprecated 3.2
+     *
+     * @param TokenHandler $handler
+     *
+     * @return Factory
+     */
     public function withTokenHandler(TokenHandler $handler): self
     {
+        trigger_error(
+            'The token handler is deprecated and will be removed in release 4.0 of this library.'
+            .' Use Firebase\Auth::createCustomToken() or Firebase\Auth::verifyIdToken() instead.',
+            E_USER_DEPRECATED
+        );
+
         $factory = clone $this;
         $factory->tokenHandler = $handler;
 
@@ -89,8 +152,11 @@ final class Factory
         $serviceAccount = $this->serviceAccount ?? $this->getServiceAccountDiscoverer()->discover();
         $databaseUri = $this->databaseUri ?? $this->getDatabaseUriFromServiceAccount($serviceAccount);
         $tokenHandler = $this->tokenHandler ?? $this->getDefaultTokenHandler($serviceAccount);
+        $tokenGenerator = new CustomTokenGenerator($serviceAccount);
+        $idTokenVerifier = new IdTokenVerifier($serviceAccount);
+        $auth = $this->apiKey ? $this->createAuth($this->apiKey, $tokenGenerator, $idTokenVerifier) : null;
 
-        return new Firebase($serviceAccount, $databaseUri, $tokenHandler);
+        return new Firebase($serviceAccount, $databaseUri, $tokenHandler, $auth);
     }
 
     private function getServiceAccountDiscoverer(): Discoverer
@@ -110,5 +176,25 @@ final class Factory
             $serviceAccount->getClientEmail(),
             $serviceAccount->getPrivateKey()
         );
+    }
+
+    private function createAuth(string $apiKey, CustomTokenGenerator $customTokenGenerator, IdTokenVerifier $idTokenVerifier): Auth
+    {
+        $client = $this->createAuthApiClient($apiKey);
+
+        return new Auth($client, $customTokenGenerator, $idTokenVerifier);
+    }
+
+    private function createAuthApiClient(string $apiKey): Firebase\Auth\ApiClient
+    {
+        $stack = HandlerStack::create();
+        $stack->push(Middleware::ensureApiKey($apiKey), 'ensure_api_key');
+
+        $httpClient = new Client([
+            'base_uri' => 'https://www.googleapis.com/identitytoolkit/v3/relyingparty/',
+            'handler' => $stack,
+        ]);
+
+        return new Firebase\Auth\ApiClient($httpClient);
     }
 }