+++ src/Firebase/Http/Middleware.php
@@ -2,6 +2,7 @@
 
 namespace Kreait\Firebase\Http;
 
+use GuzzleHttp\Psr7;
 use Psr\Http\Message\RequestInterface;
 
 class Middleware
@@ -18,7 +19,7 @@ class Middleware
                 $uri = $request->getUri();
                 $path = $uri->getPath();
 
-                if (substr($path, -5) !== '.json') {
+                if ('.json' !== substr($path, -5)) {
                     /** @noinspection ExceptionsAnnotatingAndHandlingInspection */
                     $uri = $uri->withPath($path.'.json');
                     $request = $request->withUri($uri);
@@ -41,6 +42,32 @@ class Middleware
         return function (callable $handler) use ($override) {
             return function (RequestInterface $request, array $options = []) use ($handler, $override) {
                 return $handler($override->authenticateRequest($request), $options);
+            };
+        };
+    }
+
+    /**
+     * Ensures that the API Key is present as a query parameter.
+     *
+     * @param string $apiKey
+     *
+     * @return callable
+     */
+    public static function ensureApiKey(string $apiKey): callable
+    {
+        return function (callable $handler) use ($apiKey) {
+            return function (RequestInterface $request, array $options = []) use ($handler, $apiKey) {
+                $uri = $request->getUri();
+
+                $queryParams = ['key' => $apiKey] + Psr7\parse_query($uri->getQuery());
+
+                /** @noinspection ExceptionsAnnotatingAndHandlingInspection */
+                $newUri = $uri->withQuery(Psr7\build_query($queryParams));
+
+                /** @noinspection ExceptionsAnnotatingAndHandlingInspection */
+                $request = $request->withUri($newUri);
+
+                return $handler($request, $options);
             };
         };
     }